+++ projects/ng-wizi-bulma/src/lib/date-picker/date-picker-intl.ts
@@ -1,0 +1,16 @@
+import { Injectable } from '@angular/core';
+import { Subject } from 'rxjs';
+
+@Injectable({ providedIn: 'root' })
+export class DatePickerIntl {
+  readonly changes: Subject<void> = new Subject<void>();
+  cancelLabelView = 'Cancel';
+
+  clearLabelView = 'Clear';
+
+  todayLabelView = 'Today';
+
+  nowLabelView = 'Now';
+
+  validateLabelView = 'Validate';
+}